+++ packages/local-prover/scripts/genSubtreeUpdateTestCase.ts
@@ -14,7 +14,7 @@ import {
 } from "@nocturne-xyz/sdk";
 import { WasmSubtreeUpdateProver } from "../src/subtreeUpdate";
 import { RapidsnarkSubtreeUpdateProver } from "@nocturne-xyz/subtree-updater";
-import { poseidonBN } from "@nocturne-xyz/circuit-utils";
+import { poseidonBN } from "@nocturne-xyz/crypto-utils";
 
 const ROOT_DIR = findWorkspaceRoot()!;
 const FIXTURE_PATH = path.join(ROOT_DIR, "fixtures/subtreeupdateProof.json");